refactor: create retry hint component in shared
