+++ packages/mask/src/plugins/Tips/locales/en-US.json
@@ -28,8 +28,6 @@
     "tip_add_collectibles_token_id": "Token ID",
     "tip_add_collectibles_error": "The contract address is incorrect or the collectible does not belong to you.",
     "tip_loading": "Loading",
-    "tip_load_failed": "Load failed",
-    "tip_load_retry": "Reload",
     "tip_empty_nft": "No any collectible is available to preview. Please add your collectible here.",
     "tip_disconnect_dialog_title": "Disconnect Wallet?",
     "tip_persona_sign_success": "Persona signed successfully.",
