+++ packages/shared/src/UI/components/index.ts
@@ -17,3 +17,4 @@ export * from './TokenAmountPanel'
 export * from './TokenIcon'
 export * from './WalletIcon'
 export * from './ElementAnchor'
+export * from './LoadRetry'
