+++ packages/shared/src/locales/en-US.json
@@ -27,5 +27,7 @@
     "dialog_confirm_label": "Confirm",
     "share_dialog_transaction": "Transaction",
     "share_dialog_view_on_explorer": "View on Explorer",
-    "share_dialog_transaction_confirmed": "Your transaction was confirmed!"
+    "share_dialog_transaction_confirmed": "Your transaction was confirmed!",
+    "load_failed": "Load failed",
+    "load_retry": "Reload"
 }
